Fix cart money rounding, whiten card bubbles, update team roster, olive buttons
- Rework Product/CartItem/CartRepository to carry money as BigDecimal
  instead of double, with a Money formatting helper, so subtotal + tax
  always reconciles exactly with the displayed total.
- Give the Home About Us/Our Space blurbs and every product/cart/team
  card a white background with fixed text colors, so they stay legible
  in both light and dark theme instead of blending into the page.
- Update team roster names in SampleDataProvider.
- Theme colorSecondaryContainer/colorOnSecondaryContainer olive so the
  tonal buttons (Add to Cart, cart +/-) and their ripples render olive
  instead of Material3's default lavender.

+++ app/src/main/java/com/example/jojos_cup_of_jo/model/Product.java
@@ -1,11 +1,13 @@
 package com.example.jojos_cup_of_jo.model;
 
+import java.math.BigDecimal;
+
 public class Product {
 
     private final String id;
     private final String name;
     private final String description;
-    private final double price;
+    private final BigDecimal price;
     private final ProductType type;
     private final ProductCategory category;
     private final int swatchIndex;
@@ -22,7 +24,9 @@ public class Product {
         this.id = id;
         this.name = name;
         this.description = description;
-        this.price = price;
+        // valueOf, never new BigDecimal(double): valueOf goes through Double.toString, so the
+        // literal 4.50 becomes exactly 4.5 rather than its binary expansion 4.5000000000000000277...
+        this.price = BigDecimal.valueOf(price);
         this.type = type;
         this.category = category;
         this.swatchIndex = swatchIndex;
@@ -41,7 +45,8 @@ public class Product {
         return description;
     }
 
-    public double getPrice() {
+    /** Unit price as an exact decimal. Money is never a {@code double} past this point. */
+    public BigDecimal getPrice() {
         return price;
     }
